+++ ft_date_htos.c
@@ -6,13 +6,13 @@
 /*   By: ntrancha <ntrancha@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/08/07 02:45:48 by ntrancha          #+#    #+#             */
-/*   Updated: 2015/08/07 14:43:17 by ntrancha         ###   ########.fr       */
+/*   Updated: 2015/08/11 05:53:32 by ntrancha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/date.h"
 
-int     ft_date_htos(int hour)
+long    ft_date_htos(int hour)
 {
     return (ft_date_mtos(hour * 60));
 }
